Use only the network subgraph(s) – broken!
Data union graphs seem to be gone (?). Apparently we can do DU stuff using `network` graph.
