--- src/queries/hub.ts
@@ -1,63 +1,0 @@
-import { gql } from '@apollo/client'
-
-gql`
-    fragment ProjectFields on Project {
-        id
-        domainIds
-        score
-        metadata
-        streams
-        minimumSubscriptionSeconds
-        createdAt
-        updatedAt
-        isDataUnion
-        paymentDetails {
-            domainId
-            beneficiary
-            pricingTokenAddress
-            pricePerSecond
-        }
-        subscriptions {
-            userAddress
-            endTimestamp
-        }
-        permissions {
-            userAddress
-            canBuy
-            canDelete
-            canEdit
-            canGrant
-        }
-        purchases {
-            subscriber
-            subscriptionSeconds
-            price
-            fee
-            purchasedAt
-        }
-    }
-
-    query getProject($id: ID!) {
-        project(id: $id) {
-            ...ProjectFields
-        }
-    }
-
-    query getProjects($first: Int, $skip: Int, $where: Project_filter) {
-        projects(
-            first: $first
-            skip: $skip
-            orderBy: score
-            orderDirection: desc
-            where: $where
-        ) {
-            ...ProjectFields
-        }
-    }
-
-    query getProjectsByText($first: Int, $skip: Int, $text: String!) {
-        projectSearch(first: $first, skip: $skip, text: $text) {
-            ...ProjectFields
-        }
-    }
-`
