+++ src/queries/network.ts
@@ -95,3 +95,86 @@ gql`
         }
     }
 `
+
+gql`
+    fragment ProjectFields on Project {
+        id
+        domainIds
+        score
+        metadata
+        streams
+        minimumSubscriptionSeconds
+        createdAt
+        updatedAt
+        isDataUnion
+        paymentDetails {
+            domainId
+            beneficiary
+            pricingTokenAddress
+            pricePerSecond
+        }
+        subscriptions {
+            userAddress
+            endTimestamp
+        }
+        permissions {
+            userAddress
+            canBuy
+            canDelete
+            canEdit
+            canGrant
+        }
+        purchases {
+            subscriber
+            subscriptionSeconds
+            price
+            fee
+            purchasedAt
+        }
+    }
+
+    query getProject($id: ID!) {
+        project(id: $id) {
+            ...ProjectFields
+        }
+    }
+
+    query getProjects($first: Int, $skip: Int, $where: Project_filter) {
+        projects(
+            first: $first
+            skip: $skip
+            orderBy: score
+            orderDirection: desc
+            where: $where
+        ) {
+            ...ProjectFields
+        }
+    }
+
+    query getProjectsByText($first: Int, $skip: Int, $text: String!) {
+        projectSearch(first: $first, skip: $skip, text: $text) {
+            ...ProjectFields
+        }
+    }
+
+    query getProjectSubscriptions($id: ID!) {
+        project(id: $id) {
+            subscriptions {
+                userAddress
+                endTimestamp
+            }
+        }
+    }
+
+    query getProjectForPurchase($id: ID!) {
+        project(id: $id) {
+            streams
+            paymentDetails {
+                domainId
+                beneficiary
+                pricingTokenAddress
+                pricePerSecond
+            }
+        }
+    }
+`
